Removed redundand BookingDialogBodyData
